requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.28
websocket-client>=1.5
python-dotenv>=1.0
gspread>=5.7
oauth2client>=4.1
